# Improved performance of db.models.options.Options.get_parent_list()