A similar change to #13878. My current thinking that each of these needs to review on its own merits, and therefore opened a separate PR. Happy to combine or open tickets, let me know. 

A focused benchmark script is [here](https://gist.github.com/smithdc1/fc8861434840f9a9e65a1a6e358df2e6) and I'm seeing a c.2.3x improvement. 

**Before**
```get_parent_list: Mean +- std dev: 275 us +- 5 us```

**After**
```get_parent_list: Mean +- std dev: 119 us +- 3 us```