@felixxm 2021-01-27
@smithdc1 Thanks :+1: , however I'm closing per comments in #13878. I would rather focus on improving performance of the `OrderedSet` itself.